refactor(library): use runtime-owned physicalai.config
