+++ library/src/physicalai/policies/smolvla/policy.py
@@ -15,7 +15,7 @@ from typing import TYPE_CHECKING, Any
 
 import torch
 from huggingface_hub import hf_hub_download
-from physicalai.inference.manifest import ComponentSpec
+from physicalai.config import ComponentSpec
 from safetensors.torch import load_file
 
 from physicalai.data.observation import ACTION, STATE
